move entities to 'models' folder.
