+++ models/base_entity.py
